Possible fix for updating a channel
My best working theory is that Discord returns the okay for saving the part before it actually finishes updating it on its side, so sending a message as soon as you get the okay has a chance to encounter a race condition where you send the message on a channel that doesn't really exist yet, and discord clones the channel using the old data and uses it to send the message instead. This is mitigated by waiting two seconds after saving a channel before acting on it, or possibly sending the message prior to updating the channel. I still can't say for sure exactly where the condition is happening because everything down to the cache in the library looks correct.
I suspect if you're going to send a message after you save changes to a channel you should just send the message first and then save the channel after to avoid the hassle.
I'm going to be testing this tomorrow so regardless I should have a workaround available. So far everything tells me this is not an issue with the library itself