Lax automation moderation policies (#3)
After speaking with other members of staff it has been agreed to loosen the moderation standards. Words that are merely offensive works will not have immediately harsh punishments, however words that cannot be acceptable to use even in passing (such as discriminatory words) will continue to be moderated via automation.

The duration of bans for using non-Latin characters has been reduced from 999 years to 1 hour. Harsher punishments are to be administered manually by the administrative teams as necessary.